requirements.txt
discord
pyyaml
requests